+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Better error messages for unsuccessful HTTP status codes
 - HTTP connection pooling for improved network efficiency
+- "About" menu to view project version and summary
 
 ## [1.0.1] - 2024-09-07